requirement.txt
torch
torchvision
timm
tqdm
Pillow